Inserting docker network and fixing price type

+++ src/main/resources/db/migration/V1__DDL_CREATE_STRUCTURE.sql
@@ -4,7 +4,7 @@ create table sales_order
     external_id     varchar(100) not null,
     client_document varchar(20) not null,
     product_code    varchar(50) not null,
-    price           money NOT NULL,
+    price           float NOT NULL,
     status          varchar(20) not null,
     registry_date   timestamp not null,
     last_update     timestamp not null
